Support resetting patron PINs using FOLIO APIs
Closes #739

This implements a user flow where the patron can request that an
email be sent to them that includes a token, which can be used
to reset their PIN. Symphony generates this email, but for FOLIO
we need to generate and send it ourselves, including the token.

You can view a preview of what the email looks like by visiting
/rails/mailers/reset_pins/reset_pin. It's designed to look the
same as the ones Symphony generates currently.